requirements.txt
Flask==3.1.2
gunicorn==23.0.0
asyncio==3.4.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
azure-identity==1.14.0
azure-cosmos==4.9.0
sentence-transformers==5.1.0
scikit-learn==1.3.2
aiohttp==3.10.0
gensim==4.3.3
pyodbc==5.2.0
azure-storage-blob==12.22.0